First pass at a basic character builder
README is drafted. This commit also includes
a basic attribute-->stat mapping module, a
module for deriving stat scores, and a
top-level character_builder module that should
allow for easy simulation and testing.

Code is super simple for now. I want to remove
as many 'if' blocks as possible and just have
simple computations for stats, but I am going to
hold off on optimizing until I get some balancing
data. Next priority is to factor in equipment and
level.