Laid groundwork for fully-equipped characters
This is a bit of weird commit. Were I less lazy
I'd break it up.

There are a few file renames for sanity purposes.
I made the top level builder program more logical;
the three builder methods now just return the
character and leave it to others to do stuff with
it. There are some unwritten methods referenced
in the fully_equipped_character builder method.
This represents how I intend to actually code that
functionality. It should be in the next commit.
There are also some tweaks to a few class modules.

There are also some empty json files --> 'everything'
and 'types'. Just an idea that occurred to me. More
on them later.