Simplified class / level progression
The idea of starting as some generic base
class and then moving on to a specialized
class, while not terribly unique or complex,
is nonetheless not the simplest starting
point. While my first game worked this way,
other people might prefer starting classes
to immediately confer attribute benefits.

It is pretty easy for a developer to fork
this project ad either revert or create their
own progression.