Removing proprietary concepts
These are "Daughters of Ares" concepts that
need to be separate from the open source RPG
engine. Other things may fall into this
category as I go over docs and code.

I doubt I would care much if someone went
through the commit history and copied these
ideas. It's just that they are not core concepts
of the engine.