Modified skill object
I am still unsure of this object definition.
Skills are complicated little beasts. I may
need to use more lists in places where I am
now using strings. Had to add a restrictions
field to account for actual use in combat

I added a few more skills to everything.json.
Adding them (and really any large amount of
JSON editing) is tedious, so it might be a
good idea to create a simple builder script
that just pops out the JSON I want or appends
it to everything.json or wherever. Added it
to the TODO.