Created content builder script
This seems to work fine for just adding
things to everything.json. It's kind of
janky - it doesn't format the JSON exactly
how I want and there are probably bugs
somewhere in how it handles lists or dicts
or something, but whatevz, git r done.

I should probably start writing tests or
something. Later on I'll give the script
a workout and make a shit-ton of new
content.