Character builder now allows for adding skills
The fourth version of the character builder now
produces a list of eligible skills for a
character based on their class, level, and
attributes. The user is then prompted to select
a number of skills (based on character level)
which are then added to the final character
object.

This is just a proof of concept. I need to create
more content (skills) and clean up this code. It
is nasty - silent failures and bugs out the
wazoo.