Modified skill object and deleted some JSON
Re-deleted some useless JSON files.

I altered the skill object to use lists
instead of strings in a few places (as I
had predicted I might need to do). It
is really a question of defaults; In the
vanilla version of the game engine, do I
want class skills to be restricted to a
single class or possibly multiple?

For now, the answer is "multiple."