ROADMAP update
Putting this as a separate commit because it
is a good idea. Eventually we want to make it
really easy to make complete games. Having to
manually create all of the content in the game
using the tool is, while better than editing
JSON, still not ideal.

I'll need to write a program that can a simple
CSV of content (weapons, skills, etc) and just
translate it to the big everything.json blob
I am using now.