Added new armors and fixed armor object
Armors needed to have a 'capacity' attribute.
This is an important point of difference
between individual armors, since all armors of
the same class have the same defense roll.

This commit fixes the armor object as well as
the JSON of the existing armor pieces. I also
added a few more pieces of armor using the
content tool.