Setup directory structure for engines and games
This a a big commit that does two things.

1. I have added several directories to help
enforce the concepts of 'games' and 'engines'.

Engines are top level; for now they subsume the
Python files (this may have to change in order
to make engines creatable). Engines have a
'games' subdirectory, which consists of games
based on a given engine.

Games are defined by their JSON blobs - the
content, the skills, classes, leveling, gear,
etc.

For now there is only the default engine and
the default game.

2. I am using a single content class to handle
getting data from everything.json. The only
exception is addContent (and later editContent)
which have to actually change the everything.json
file.