Added classes and characters to content builder
There are some other minor tweaks to this commit
that I am too lazy to make into their own thing.

1. I added a "tag" attribute to characters. This
is how I am going to deal with enemy characters,
including monsters. All of a game's enemies will
just be character objects.

2. I delted levels.json. everything.json can
handle it.

3. (the main thing) the content builder can now
make classes and character objects. This only
required a template update and a few tweaks to
methods in addContent.py

addContent.py now also uses Content() objects.
I will have to tweak how it is instantiated
so that it can take in the name of a game.