First step in convering engine to JSON
I have a way forward for converting
attStats.py into JSON and only having
high-level attribute and stat functions.
What's nice is that the JSON should be
easy to generate programtically, and yet
it is fairly flexible in terms of what it
will let engine builders do with stat and
attribute relationships.