More progress converting engine to json
Editing JSON sure is tedious...

A number of decisions are implied in this
commit. One is the explicit mappings of
all possible attribute values to specific
modifiers. Even though it is annoying to
code by hand, it is easy to produce from
a template, which is the ultimate goal

What's more, embedding mathematical
expressions in JSON is a bit dicey. There
is a Python library that can parse expressions
in string form, but I can't assume it will be
reliable for all of the weird formulae that
users might come up with.

It's easier to just have users explicitly
map all attribute values to an output. This
is what most rulebooks already do. The only
disadvantage is that it makes it difficult
to use very large numbers for attributes and
stats. Addressing this issue is an optimization
for another time.

The next commit should see attMappings and
most of attStats made obsolete.