Converted attribute / stats engine to JSON
This is an important commit. This will serve
as the model for how to convert the Python
modules to JSON.

This commit completes the work begun in the
last and therefore removes the need for
attMappings.py and most of attStats. Changes
were made to the builder and character
modules to reflect the engine change and a
few other simplifcations (fewer stats,
shorter names).

The transition to a fully JSON engine will
be comprised of four stages that follow the
core methods of builder.py:

1. Naked Level Zero Characters

This is complete. The data needed to define
a character given only attribute values is
all in JSON, hence the first builder method
in the builder module now works with the
engine.json file.

2. Leveled naked characters with classes
3. Characters with classes, levels, and gear
4. All of the above + skills

At some point I may add races. Other things
that may be necessary:

1. Add concept of race and gender
2. Add types to stats for analytical purposes

That's all for now.