Created template for character constructs
This commit begins the work of making character.py
obsolete. The level_zero_character_builder method
now uses the engine.json template to construct
character objects (which are now just dictionaries.
It is unclear if there is a need to make Python
objects out of them).

As I adapt the later builder methods to use this
template I will eventually come to a place where I
can just delete character.py. This will be a big step.