Restructuring. Lots of restructuring.
This commit sets the stage for a few things.

For one there are some name changes.
"everything.json" is now "content.json." This
is how game content will be defined. Game
content is by default distinct from the engine,
however the two can be combined by using the
"CONTENT" construct in the engine.json file.

A developer can embed ALL of their game content -
their classes, weapons, skills, etc. - into the
engine's CONTENT construct and not even use the
content.json file. Most likely, that content.json
file will still have to exist in a subdirectory
with the game name if only to enforce the convention
and add flexibility.

There are some other file renames. Among them is
"engine.py" which will be the high level module
for working with engine.json. It should eventually
be flexible enough to parse and work with custom
constructs but for now it will just look for certain
conventional constructs, like "LEVELS" and "EQUIPMENT."