Bugfix and other tweaks
This commit resolves the bug where class
mods were not being applied at each
level up. This was caused by a semantic
error in a for loop that caused only part
of each level object to be applied to the
character.

I moved around some other data as well. I
realize now that I need a consistently
enforced pattern of accessing data in the
content.json and engine.json files.

The convention should be that whenever a
method needs a construct or engine-level
information, it should check engine.json.
If it needs information about content, it
should first check content.json and if
it does not find it or if content.json
does not exist, it then checks the
engine's CONTENT construct for the data.

This is roughly what I want as it keeps
things flexible.