Deleted items.py
Characters can now equip weapons and
have an inventory of gear without the
need of the items.py module.

This is a pretty quick first pass and
probably buggy as hell, but it'll do
for now. I'll revisit and polish it
later.

We are on the cusp of having a full
engine written only in machine cretable
JSON. There is a lot of work to do around
the overall data structure, and skills are
probably going to be a BITCH to code, but
at least we're making progress.