Cleared the way for skillz
There was a bug in the equip method. A typo.

I also need to revisit this game's notion of
"combat stats," because I don't like the way
the data is organized now.

The 'defense' attribute is nested under the
charac["equipment"] dictionary. I don't know
if that is ideal. The way armor-class-based
defense is laid out is also janky but seems
to at least work.

Still, I like this commit, because it kills
two more uselss Python modules, and sets a
path for the skill commit to come.