Skills are now part of the JSON engine
This is an important accomplishment.

With this commit, all of the core constructs
are now part of a trivially generated JSON
blob.

Still a lot of bugs and wonkiness in this code.
I need to make engine.py smarter. 'Engine
Builder' will need to know how to deal with
custom constructs and unconventional game
design choices.

Still, this is a pretty cool place to be.