Added random attribute allocation for builder
This commit adds functionality to builder to
allow for random allocation of attrtibute
bonuses for level ups.

This also includes a test for the custom
level character builder. It's actually a
crappy test, as it only verifies that the
MaxHP bonuses were applied. I need to think
of a smart way to fix the player-allocated
bonuses.