Updated addContent to work with game content.json
This is kind of a big deal.

Game content can now be easily added to the game
content.json using the addContent module. No
need for directly editing the JSON. With some
polish, we can build this out into a nice way
to add lots of content to games quickly.