Updated editContent to work with game content.json
This commit also fixes a bug I missed with the
addContent module. It was getting the content
dictionary one layer too low and therefore not
creating proper objects.

Basic content editing is possible with this module.
It still requires first selecting a type. Ideally
it would be smart enough / search fast enough to
just take in a name and then search all content in
content.json for a match. The files should be small
enough for that always to be very quick.

More importantly, with the ability to quickly add
and edit content, we can start to work on the real
Game Builder module.

Exciting times indeed.