Updated engine.py and tests
Everything now works with the new engine.json.
This is a proof of concept commit. If I stick
with this structure for engine files, I will
need to make a number of improvements to
engine.py. More precisely, engine.py will
have to be smart enough to know how to make
a character without knowing which Schemes
to expect in the engine.json file.