Shuffling things around for packaging
This commit does a few small things that push
Otrera toward proper Python packaging.

1. Adds a placeholder yaml. That will be used for
configuring things like where engine and game
data will be saved on the user's box.

2. Moves all the Python code to the top level
'otrera' directory.

3. Makes content.py a bit smarter. It will still
default to the default engine and content.json, but
it can take in paths for specific engines and games.

4. Deleted template.py. Pretty sure that's no longer
used.

It's so nice having an integration test to make sure
things work!