Completed first iteration of Game Builder
This includes an update to the otrera
executable. Pretty happy with this commit
overall. You can create a new game and then
add content to it.

I need to update the editContent module and
then add some kind of test. Going to need a
few commits to really polish the CLI.