Restructed content.json for new data model
This commit breaks tests, so hopefully I won't
dilly-dally on finishing the bigger task.

The goal here is to make content.json reflect the
Construct --> Scheme --> Content data model. This
required shifting around a lot of keys and rewriting
much of content.json by hand.

The goal is to have content.json be almost entirely
root keys that refer to pieces of content. The only
exceptions are the "List" and "Types" keys, which
are embedded in content.json for a few convenience
methods (more on that later)

This commit accomplishes this first goal. Next is the
likely more irksome task of trying to make engine.py
work with this new model.