Added clone method and fixed engine hardcoding
Game Builder had "default" hardcoded into it
for the engine name in a number of places. Now
it will take an engine name as an argument and
default to 'default'.

I also added a clone method that will allow
you to copy an existing game and just give it a
new name.