Next big step in encoding Pathfinder
This commit sees the basic skeleton
of the game encoded. One can't say
that the engine is done since all
of the rules aren't written out.
The structure of the game is there
though, in the JSON.

A bunch of content is done too. All of
the beginner box classes, races, and
equipment are done. Next commit should
include spells / skills / feats. Then
another commit can be enemies. Text
for rules and descriptions will come
last.