Renamed some files, changed priorities
This commit properly names the Pathfinder
directory; I want to make a distinction
between the complete game and the
beginner box.

This commit also represents a switch in
priorities. My last commit (mistakenly)
predicted that the next series of commits
would all be content encoding for
Pathfinder. Circumstances have made it such
that this is not the best use of time.